Add tag-summary pages. These are very slow.
Possibly these should just be occassionally pre computed at the end of
a semester, as they are slow.

These show the piechart completion summaries, both overall and
seperately by instrument, weatherband and PI country, based on the
initial allocations. It requires a Hedwig
'M_<semestername>_P_info.json' file to be present to get the initial
allocations : these files should not be committed to git as they
contain proprietary information.

It also provides a sort-able table of allocation and observational info.