summary: improve msb information.
Now show a table of the msbs below the charts. Also fix some bugs so
that the weather band is correctly calculated, make the tables and the
whole weather band sections able to be hidden on click, and add a
warning that moving targets aren't shown in the observability plots.