summary: handle observations with no JCMT tau in detailed project view
Currently these are shown as a warning. Really these should be moved
to some sort of informational section, as it can't be changed but just
affects the plots of breakdown of observations by weather band. Should
probably ignore JUNK (and bad?) observations too in the future.

M16AL001 has a (JUNK) observation with no wvmtau, which is how this
was found.